bmt-to-quarkus: Moved model class to its own package, added package info attaching it to 'primary' PersistenceUnit
