+++ bmt-to-quarkus/src/main/java/org/jboss/as/quickstarts/bmt/model/KVPair.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.as.quickstarts.bmt;
+package org.jboss.as.quickstarts.bmt.model;
 
 import java.io.Serializable;
