+++ bmt-to-quarkus/src/main/java/org/jboss/as/quickstarts/bmt/model/package-info.java
@@ -1,0 +1,4 @@
+@PersistenceUnit("primary")
+package org.jboss.as.quickstarts.bmt.model;
+
+import io.quarkus.hibernate.orm.PersistenceUnit;
